GIX-1212: Rename file
