+++ frontend/src/lib/modals/proposals/NnsProposalsFilterModal.svelte
