+++ frontend/src/tests/lib/modals/proposals/NnsProposalsFilterModal.spec.ts
@@ -3,7 +3,7 @@
  */
 
 import { DEFAULT_PROPOSALS_FILTERS } from "$lib/constants/proposals.constants";
-import ProposalsFilterModal from "$lib/modals/proposals/ProposalsFilterModal.svelte";
+import ProposalsFilterModal from "$lib/modals/proposals/NnsProposalsFilterModal.svelte";
 import { proposalsFiltersStore } from "$lib/stores/proposals.store";
 import type { ProposalsFilterModalProps } from "$lib/types/proposals";
 import { enumKeys } from "$lib/utils/enum.utils";
